+++ app/util/util_param.py
@@ -13,6 +13,7 @@ def get_init_info ( request:Request, post:object = None, only_params:bool = Fals
     params =  get_predefined_params ( request, post )
     params["@ip"] = util_library.get_client_ip(request)
     params["@lang"] = const.CONF["locale"]["lang"]
+    params["@ui"] = request.cookies.get("ui")
 
     if post :
         for key, val in post.items():
@@ -57,6 +58,31 @@ def get_init_info ( request:Request, post:object = None, only_params:bool = Fals
 
     return panel, panel_json, params
 
+def get_panel_defaults ( panel_json, params ) :
+    if "chart" in panel_json :
+        if "defaults" in panel_json["chart"] :
+            for key, val in panel_json["chart"]["defaults"].items() :
+                if key == "@date" : key = ".date"
+                if key not in params or params[key] == "" :
+                    params[key] = val
+                    if isinstance(val, str) and len(val) > 0 and val[0] == "@" :
+                        default_query = panel_json["chart"]["conditions"][key]["values"]["query"]
+                        default_datasource = panel_json["datasource"]
+                        if "datasource" in panel_json["chart"]["conditions"][key]["values"] :
+                            default_datasource = panel_json["chart"]["conditions"][key]["values"]["datasource"]
+                        default_values = util_db.select_db(default_datasource, default_query, params)["data"]
+
+                        if val == "@first" :
+                            val_key = list(default_values[0])
+                            val = default_values[0][val_key[-1]]
+                        elif val == "@last" :
+                            val_key = list(default_values[-1])
+                            val = default_values[-1][val_key[-1]]
+
+                        params[key] = val
+
+    params = get_predefined_time_params(params)
+    return params
         
 def get_predefined_params ( request:Request, post:object = None ) :